ffi(bridge): split sync DTOs into sync/dto.rs (pre-500 refactor)
