+++ ffi/secretary-ffi-bridge/src/sync/mod.rs
@@ -3,8 +3,10 @@
 //! the desktop consumes them as a Rust crate; projecting them onto uniffi/pyo3
 //! is deferred to issue #187 (mirrors #167 for the contacts/revoke functions).
 
+pub mod dto;
 pub mod orchestration;
 pub mod status;
 
-pub use orchestration::{sync_vault, CollisionDto, SyncOutcomeDto, VetoDecisionDto, VetoDto};
+pub use dto::{CollisionDto, SyncOutcomeDto, VetoDecisionDto, VetoDto};
+pub use orchestration::sync_vault;
 pub use status::{sync_status, DeviceClockDto, SyncStatusDto};
